delete old metrics
also take my helper methods from old metricshandler
