+++ src/main/java/com/griefprevention/metrics/MetricsHandler.java
@@ -2,6 +2,8 @@ package com.griefprevention.metrics;
 
 import me.ryanhamshire.GriefPrevention.GriefPrevention;
 
+import java.util.concurrent.Callable;
+
 /**
  * Created on 3/2/2024.
  *
@@ -17,8 +19,25 @@ public class MetricsHandler
 
         try
         {
-            metrics.addCustomChart(new Metrics.SimplePie("bukkit_implementation", () -> plugin.getServer().getVersion().split("-")[1]));
+            addSimplePie("bukkit_implementation", plugin.getServer().getVersion().split("-")[1]);
         }
         catch (Throwable ignored) {}
+    }
+
+    private void addSimplePie(String id, boolean value)
+    {
+        addSimplePie(id, Boolean.toString(value));
+    }
+
+    private void addSimplePie(String id, String value)
+    {
+        metrics.addCustomChart(new Metrics.SimplePie(id, new Callable<String>()
+        {
+            @Override
+            public String call() throws Exception
+            {
+                return value;
+            }
+        }));
     }
 }
